feat: 타입들 namespace를 사용해서 모듈 단위로 묶음
