+++ src/pages/MyAccount/myAccount.d.ts
@@ -1,0 +1,3 @@
+declare namespace MyAccount {
+
+}
